@@ -15,10 +15,9 @@
     "es6-promise": "^4.1.0",
     "isomorphic-fetch": "^2.2.1",
     "lodash": "^4.17.4",
-    "react": "^15.5.4",
-    "react-dom": "^15.5.4",
+    "react": "16.5.2",
+    "react-dom": "16.5.2",
     "react-hot-loader": "^3.0.0-beta.7",
-    "react-youtube-playlist": "CompanionLabs/react-youtube-playlist",
     "sass-loader": "^6.0.5",
     "style-loader": "^0.18.1"
   },
