@@ -17,8 +17,8 @@
   "author": "",
   "license": "ISC",
   "peerDependencies": {
-    "react": "^15.5.4",
-    "react-dom": "^15.5.4"
+    "react": "16.5.2",
+    "react-dom": "16.5.2"
   },
   "dependencies": {
     "babel-polyfill": "^6.23.0",
@@ -26,13 +26,12 @@
     "es6-promise": "^4.1.0",
     "isomorphic-fetch": "^2.2.1",
     "jquery": "^3.2.1",
-    "react": "^15.5.4",
+    "react": "16.5.2",
     "react-bootstrap": "^0.31.0",
-    "react-dom": "^15.5.4",
+    "react-dom": "16.5.2",
     "react-proptypes": "^1.0.0"
   },
   "devDependencies": {
-    "@companionlabs/gulp-common": "^0.4.8",
     "babel-cli": "^6.26.0",
     "babel-core": "^6.24.1",
     "babel-plugin-module-resolver": "^2.7.1",
