@@ -10,8 +10,8 @@ const App = () => {
       api_key="AIzaSyBRKbR9hS9678OzzL0Ju7XjLZPISRpKnwY"
       playlist_id="PLCP9GgzdebSIUq_GmmRU_OdmDhSLBC5lP"
       show_thumbnails
-      width="800px"
-      height="600px"
+      width="100%"
+      height="100%"
     />
   )
 }
